splitting TestingDAQProvider and RSAAcquisitionProvider into their own modules
